libnm: tigthen NMTeamLinkWatcher structure and embed strings
NMTeamLinkWatcher is a simple and immutable data structure.

It is a good thing that it's immutable today, and as that is a desirable
property, we should also aim to have the NMTeamLinkWatcher immutable in
the future.

As such, we can pack the NMTeamLinkWatcher tighther. Instead of
allocating separate strings, append them after the memory of the
structure.